requirements.txt
flask==2.3.3
werkzeug==2.3.7
Pillow==10.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
gevent==23.9.1
Flask-Caching==2.1.0
